+++ R/controller.R
@@ -8,6 +8,8 @@ install_phantomjs(force = TRUE)
 #ud_model <- udpipe_load_model(ud_model$file_model)
 ud_model <- udpipe_load_model('../tools/data/english-ewt-ud-2.5-191206.udpipe')
 #-------------------------------------------------------------------------------------------
+correction.tidy()
+Sys.time()
 source('generate list of UPIs.R')
 Sys.time()
 source('generate wordcloud for each UPI.R')